build fix

+++ inc/FSM/Libs/Context.h
@@ -438,7 +438,7 @@ void Context<T>::handleEvent(_pulse event)
     Topic event_code = (Topic)event.code;
     int event_value = event.value.sival_int;
     data->event_payload = event_value;
-    data->event_topic = event_code;
+    data->event_topic = (int8_t) event_code;
     switch(event_code) {
         case Topic::INTERRUPT:
             newState = handleInterrupt(event_value);
